Detect host machine type and set CFLAGS accordingly.
The makefile was defaulting to building with assembly for x86,
meaning it wouldn't build out of the box on an arm host. This
patch tries to be clever, and enables different flags based
on the output of `uname -m`. There isn't currently any arm
assembly, but we define OD_ARMASM as an example of how to add
other architectures.