Fix MacOS: Copy the cpuid macro from libtheora.
The cpuid macro tries to preserve ebx (used for -fPIC) by
exchanging it with whatever location the caller's 'ebx'
variable is using on either side of the cpuid operation.

However, this has a bug on x86_64: round-tripping through
the 32-bit ebx will clear the upper 32 bits of any 64 bit
value being held in that location. In particular, the
optimizer in llvm-gcc and clang 3 from Apple XCode 4.2.1
tries to reuse a copy of the _state pointer which was
held in that register at the time of the call, and
subsequently segfaults dereferencing the truncated
address (-O1 and higher).

Contrariwise, if we add the register to the clobber
list, gcc cannot satisfy the dependencies on x86_32.

Instead, we borrow the version of the cpuid macro
from the libtheora source, which uses a different
version on x86_32 and x86_64, the later without
the ebx specified explictly for the _ebx macro
argument.