Suppress string length warnings.
ISO C90 as a length limit of 509 on string literals, and
some versions of gcc warn about this with -std=c80.

While we generally want to compile under C89, we (currently)
use long strings in two places: __asm__ blocks, and usage
text in the example programmes. The former is an extension
anyway. The later isn't a critical component and breaking
up the text hurts readability, so we choose to suppress the
warning for now.