Fix make distcheck.
Having a nested directory in top-level automake SUBDIRS caused
the directory to be build, but not dist-ed, so the 'make distcheck'
test would fail because it couldn't find src/tests/Makefile.in.

Instead, recurse according the the directory tree. We have to
put the current directory first in src/Makefile.am to ensure
it's built before the tests when depend on them.