Move jenkins build scripts into the tree.
This just does the unix and autotools builds, scan-build and
coverage still need to be fixed up and moved.

Having these in the tree makes the tests more transparent,
and something which can be run locally, instead of having
that part of the configuration tied up in the jenkins database.