Require automake 1.11 again.
We're not using 1.11-specific features, but 'distcheck' fails
under 1.10 with 'files left in build after distclean' because
it doesn't generate clean rules for pathed noinst targets.

That in itself isn't serious (except for producing public dist
packages) but we only lowerd the version requirement because
of complaints from developers with system automake 1.10 on
MacOS X. As near as I can tell, this is an unsupported version
installed by Xcode prior to version 4.3 and not updated since.

More recent versions don't include autotools at all and developers
must install manually or from a package system like homebrew if they
want to build out of the repository. It's better to encourage people
with crufty mac systems to upgrade.