Adds prediction input to block size switching code
Seems to be working, but will probably need lots of tuning. Also
fixes a bug where we were using img_stats instead of psy_stats for
masking. Also had to add a psy floor at 0 for cases where the residual
was small (not sure if it's the right thing to do).

Conflicts:
	src/encode.c