Check for libm and explicitly link it everywhere it is needed
The default for newer versions of ld is to enable --no-add-needed
(renamed to --no-copy-dt-needed-entries in later versions) which
means we need to explicitly link all the libraries we need, we
don't get them automatically from other dependencies anymore.

But that's not exactly the problem here.  The gold linker is even
more picky about this, and fails in some cases where GNU ld still
doesn't.  This patch fixes the build with gold.