Separate the tests for the two different doc types
There's no reason to not build any of them if the tools for only one
are available, they are independent at this stage.

This probably also needs to test for pdflatex, and could report what
will be enabled in a bit more detail in the final summary, but don't
get too carried with this until we know it's not all going to change.

Use && instead of 'test -a', the latter is a known portability trap.