More portable wall time
Expands on a patch from Eric L. Camachat for FreeBSD which switched to
using gettimeofday, to also test whether it or ftime is available.
The ftime function is deprecated by POSIX and has moved to libcompat
on FreeBSD, but gettimeofday is not available on Windows, so wrap them
both in our own wall time request function.