Add od_fpu_restore() to clear out FPU state after MMX/SSE
Add od_fpu_restore() to clear out FPU state after MMX/SSE use when x86 assembly
is enabled.  This doesn't go quite as far as the Theora implementation where the
'call' is a define. It just uses the state opt_vtable, and I added the relevant
code to state.c and x86/x86state.c.  It's only needed in the encoder right now,
might not be the case in the future.

Adds one call to od_fpu_restore() at the end of od_mv_est(); this is the only
place it's needed right now as far as I know.