Scripts to run a randomized comparison test.
This is very rough initial work.
The goal is to be able to test multiple versions of the same image
 against a reference, and randomize the order the images are
 presented as well as the order each version is presented.
These scripts produce a bare-bones webpage and a set of numbered
 images in a directory which can be tarred up and distributed, along
 with a key for descrambling the numbers, and a script that can
 process rankings against the key to give unscrambled totals.

There is currently no means to enter the test results in the web page.
We should probably be looking at quality measured on some scale
 instead of just rankings.
There is currently no way to declare a tie.
The page itself should probably be split into multiple pages (one per
 set of images) with keybindings for easy navigation.