Do not allow splitting a skipped region
There are cases where a split would cause all blocks to be skipped but not
splitting causes the block not to be skipped. In those situations, the block
size RDO metric (which can't perfectly match the quantization RDO) will
sometimes prefer skipping the split blocks, which is silly. We should
never split in this situation, although we can choose whether to skip the
unsplit block or code it. Right now we choose to code it since 1) it's simpler
and 2) the quantizer RDO is probably more trustworthy for now.