Change daala_decode_header_in to simplify usage.
Previously daala_decode_header_in returned 0 once it encountered a video
packet, but this required packet peeking since you'd always read one
packet too far while processing headers. Now the funtion returns the
remaining headers to read, and 0 when processing the final header.

This means we can't send arbitrary numbers of headers now, but this
functionality was never used in Theora; we don't expect to use it in
Daala either.