Add 'f' for fullscreen to player_example.
SDL will scale the image in the gpu, so this isn't ideal for
diagnostic interpretation, but is nice for display, especially
for higher resolution videos which may not fit on a single screen
in a decorated window.

This works better than SDL_WM_ToggleFullScreen() in SDL 1.2,
which tries to change the output video resolution. However
there are still some issues, e.g. gnome-shell minimizes
the window if it loses focus.

Signed-off-by: Thomas Daede <daede003@umn.edu>