Move upsample line buffer from state to encoder context
This got missed in the earlier buffer movement; the upsampling line
buffer is only used in upsample8, and so, only conditionally in
encode. This patch moves it to the encoder ctx.

In addition, the patch eliminates tools/upsample's dependency on using
the line buffer from od_state as it's not there anymore.