pvq: remove od_apply_qm
Instead, apply this scaling at the PVQ level. Fixes #135.

subset1:

         RATE (%)  DSNR (dB)
    PSNR -0.18631  0.00820
 PSNRHVS -0.14242  0.01004
    SSIM -0.17364  0.00642
FASTSSIM -0.20725  0.00703

ntt-short-1:

         RATE (%)  DSNR (dB)
    PSNR 0.00592  -0.00020
 PSNRHVS 0.03283  -0.00160
    SSIM -0.06473  0.00156
FASTSSIM 0.06960  -0.00195

Note that when FPR is turned on, we will see a larger regression
on FastSSIM on video due to this patch. This is because this patch has less
noise on P-frames, which hurts FastSSIM.

This patch also helps chroma-from-luma, since before we were predicting
blocks with od_apply_qm applied to them from neighbouring blocks which
had it unapplied.