Commit 16-bit version of SAD and SATD calculations toward enabling FPR
This patch adds the trivial 16-bit C entensions (no SSE) to SAD
calculation for supporting full precision (double-byte) references.

It also refactors the code flow (though not the actual math) used for
SATD so that the bulk of the code can be reused between 8 and 16
bit. The code is slightly slower (~10%) when using -O2, but up to 55%
faster using -O3.