Get rid of the bilinear filter
Its enhancement effect has been almost completely eliminated by a combination
of the deringing filter and 64x64 DCTs so that right now it's mostly a
loss.

subset1 results: low = 0.05-0.1, medium = 0.1-0.2, high = 0.2-0.5
          LOW (%)  MEDIUM (%) HIGH (%)
    PSNR -0.497171 -0.175219 -0.060687
 PSNRHVS -0.366111 -0.160004 -0.095976
    SSIM -0.683048 -0.336761 -0.128636
FASTSSIM -3.630780 -2.687704 -1.982495

No ntt-short1 since AWCY is broken, but the bilinear filter was only enabled
on keyframes anyway (which was already a problem).