Fix some additional overflows in the 64x64 DCT.
Test #350 could overflow even with 4-point lapping.
Tests #200 and #242 could only overflow with 8-point lapping.

Again, this has exactly zero impact on measured MSE in dcttest.
There is a very small impact on the reported biases, but the first
 several significant figures of each one do not change.

Changes in metrics on ntt-short-1 seem to be mostly noise:
          RATE (%)  DSNR (dB)
    PSNR  0.00470  -0.00014
 PSNRHVS  0.00603  -0.00028
    SSIM  0.00379  -0.00009
FASTSSIM -0.06955   0.00187