Alter golden frame spacing calculation, remove a hopefully dead frame type mutation
Land issue 1318:

A previous unstated invariant was that I-frames are also golden.  This
changed at some point, and the resulting code can now generate a
golden frame interval wider than a GOP, so a case was added to force a
frame golden when no reference was already golden, usually the P-frame
after an I-frame in the closed-GOP case.  Go back to making I-frames
golden.

A second case forces an I frame when there is no reference frame.
This case should be impossible.  Remove the type mutation and add an
assert to guard the case.