Motion-Compensated Block Overlay
Reused the od_mv_grid_pt struct in the analyzer to get motion vector data from
the decoder. The number of blocks in a frame (both vertical and horizontal)
needed to be added to the decoder in order to reduce dependencies of the
analyzer. The shortcut-key crtl-M was chosen for the word Motion. This overlay
will also include Motion Vector in a subsequent commit.

Signed-off-by: Nathan E. Egge <negge@dgql.org>