Reduce RAM usage of dcttest by converting basis storage from doubles to bytes.
dynamic_range() stores the basis functions by passing impulses to the
transforms, but the range checks later only care about the polarity of
the response.

Instead of storing the basis as doubles in basis2, store just -1, 0, or 1 in
basis_pol. This reduces RAM usage from 626MB to 168MB, while yielding identical
output.